address review comments refactor EnvironmentWatcher
