+++ pom.xml
@@ -68,7 +68,6 @@
         <java.level>8</java.level>
         <no-test-jar>false</no-test-jar>
         <groovy-cps.version>1.32</groovy-cps.version>
-        <pipeline-stage-step.version>2.3</pipeline-stage-step.version>
     </properties>
     <dependencyManagement>
         <dependencies>
@@ -257,13 +256,8 @@
         <dependency>
             <groupId>org.jenkins-ci.plugins</groupId>
             <artifactId>pipeline-stage-step</artifactId>
-            <version>${pipeline-stage-step.version}</version>
+            <version>2.3</version>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.jenkins-ci.plugins</groupId>
-            <artifactId>pipeline-stage-step</artifactId>
-            <version>${pipeline-stage-step.version}</version>
         </dependency>
     </dependencies>
     <build>
