+++ src/main/java/org/jenkinsci/plugins/workflow/cps/InterpolatedSecretsDetector.java
@@ -3,28 +3,27 @@ package org.jenkinsci.plugins.workflow.cps;
 import hudson.EnvVars;
 import hudson.model.Run;
 import hudson.model.TaskListener;
-import org.jenkinsci.plugins.workflow.cps.view.EnvironmentWatcherRunReport;
+import org.jenkinsci.plugins.workflow.cps.view.InterpolatedSecretsDetectorReportAction;
 import org.jenkinsci.plugins.workflow.flow.FlowExecutionOwner;
 import org.jenkinsci.plugins.workflow.steps.EnvironmentExpander;
 
 import javax.annotation.Nonnull;
 import java.io.IOException;
-import java.io.Serializable;
 import java.util.List;
 import java.util.Set;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
-public class EnvironmentWatcher implements Serializable {
+public class InterpolatedSecretsDetector {
     private EnvVars envVars;
     private Set<String> watchedVars;
     private List<String> scanResults;
-    private static EnvironmentWatcherRunReport runReport;
+    private static InterpolatedSecretsDetectorReportAction runReport;
 
-    private static final Logger LOGGER = Logger.getLogger(EnvironmentWatcher.class.getName());
+    private static final Logger LOGGER = Logger.getLogger(InterpolatedSecretsDetector.class.getName());
 
-    public static EnvironmentWatcher of(CpsStepContext context, CpsFlowExecution exec) {
+    public static InterpolatedSecretsDetector of(CpsStepContext context, CpsFlowExecution exec) {
         try {
             EnvVars contextEnvVars = context.get(EnvVars.class);
             EnvironmentExpander contextExpander = context.get(EnvironmentExpander.class);
@@ -32,20 +31,20 @@ public class EnvironmentWatcher implements Serializable {
                 if (runReport == null) {
                     FlowExecutionOwner owner = exec.getOwner();
                     if (owner != null && owner.getExecutable() instanceof Run) {
-                        runReport = ((Run) owner.getExecutable()).getAction(EnvironmentWatcherRunReport.class);
+                        runReport = ((Run) owner.getExecutable()).getAction(InterpolatedSecretsDetectorReportAction.class);
                     }
                 }
                 if (runReport != null) {
-                    return new EnvironmentWatcher(contextEnvVars, contextExpander);
+                    return new InterpolatedSecretsDetector(contextEnvVars, contextExpander);
                 }
             }
         } catch (InterruptedException | IOException e) {
-            LOGGER.log(Level.FINE, "Unable to create EnvironmentWatcher instance.\n" + e.getMessage());
+            LOGGER.log(Level.FINE, "Unable to create EnvironmentWatcher instance.");
         }
         return null;
     }
 
-    public EnvironmentWatcher(@Nonnull EnvVars envVars, @Nonnull EnvironmentExpander expander) {
+    public InterpolatedSecretsDetector(@Nonnull EnvVars envVars, @Nonnull EnvironmentExpander expander) {
         this.envVars = envVars;
         watchedVars = expander.getSensitiveVars();
     }
